curl_ngtcp2 handshake timeout should be equal to --connect-timeout
Default timeout is hardcoded (10 seconds) and doesn't respect --connect-timeout parameter. In some cases 10 seconds can be not enough or too long to "establish a connection". Moreover the non-working --connect-timeout parameter for http3 is confusing. This change makes the handshake timeout equal to --connect-timeout, if it's set.
Discussion is here https://github.com/curl/curl/discussions/18427