tool_getparam: warn on all unicode 0xe2 0x80 prefixes
If a string argument is expected and the first two bytes are 0xe2 ex80
that's enough for curl to warn.

Previously we tried to detect and warn only for the unicode double
quote, but users might use single quotes, other quotes or even lead the
argument with one of the "zero widths" characters. This is an attempt to
detect many of those. Without triggering for "normal" IDN hostnames.