quic: fix min TSL version handling
When switching to TSLv1.2 as default in 9d8998c99408e1adf8eba629fad9f87b3235bdfa, this
led to an explicit setting of 1.2 on QUIC connections when using quictls, overriding
the already set min version of 1.3.

This leads to a ClientHello with TLS 1.2+1.3 offered on a QUIC connect which is rejected
by the Caddy server. Using ngtcp2 with OpenSSL 3.5+, GnuTLS, AWS-LC is not affected.

refs #18518