RIP pthread_cancel
Remove use of pthread_cancel in asnyc threaded resolving. While
there are system where this works, others might leak to resource
leakage (memory, files, etc.). The popular nsswitch is one example
where resolve code can be dragged in that is not prepared.

The overall promise and mechanism of pthread_cancel() is just
too brittle and the historcal design of getaddrinfo() continues
to haunt us.

refs #18532