cfilter, unlink and discard
Rewrite the code that removes a filter from the connection and
discards it. Always look at the connection, otherwise it will not
work of the filter is at the top of the chain.

Change QUIC filter setup code to always tear down the chain in
construction when an error occured.

HTTP proxy, do not remove the h1/h2 sub filter on close. Leave
it to be discarded with the connection. Avoids keeping an additional
pointer that might become dangling.

Triggered by a reported on a code bug in discard method.

reported-by: Joshua Rogers