dependabot: ignore patch releases for the codeql action
It's updated quite often:
https://github.com/github/codeql-action/releases

This has the downside that we're always using the first
patch release, and never any bugfix releases.

Ideally I'd only bump to the latest patch release of
the previous major-minor release.

Note that mend/renovate is also bumping actions, so if
that one misses it, or we cancel it, dependabot will kick
in a try bumping it also.

So perhaps we should ignore all actions bumps from
dependabot.