This change adds support for several extended AWS SIGv4 options:
A new --aws-sigv4-mode option is added to select either the (default)
header authentication mechanism, or the query string authenitcation
mechanism.
* https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_sigv-authentication-methods.html

A new --aws-sigv4-algorithm option is added to select either the
(default) HMAC-SHA256 signature algorithm used for standard single
region SIGv4 signature, or the the ECDSA-P256-SHA256 signature algorithm
used for multi-region SIGv4A signatures.
* https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_sigv-create-signed-request.html

A new --aws-sigv4-signedheaders option is added to select a specific set
of headers to include in the signature.
* https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_sigv-signing-elements.html

The change also includes a few smaller updates:

The --user parameter is updated to support optionaly specifing a
security token. This is useful for querystring signing mode, where the
parameter needs additional encoding that would be hard for users to do.
(This security token can also still be specified explicitly in an
X-Amz-Security-Token header instead when header signing mode)

This includes an update to allow the X-Amz-Date or X-Amz-Date headers to
control the date used for signing requests. This is useful when testing
different scenarios with SIGv4 or SIGv4A.