transfer: fix retry for empty downloads on reuse
When a reused connection did transfer 0 bytes, it assumed the
transfer had failed and needed a retry. Add a check for
data->red.done, so we can successfully accept the transfer of
a 0-length file via SFTP.

Add test case 1583 to verfiy.

Fix SFTP disconnect debug trace when there was nothing to
disconnect (like when reusing a connection).

refs #19165