scp/sftp: fix disconnect
When a SCP/SFTP connection calls the protocol handler disconnect, it
required the connections *and* the easy handles SSH meta data to be
present. When the disconnect is called with an admin handle, the easy
meta data is not present, which prevented the shutdown to run.

The easy meta data is however not necessary to run the shutdown
state machine. Calling it with a NULL `sshp` is fine. To avoid
any mixups, check `sshp` in state operations that need it.

refs #19293