runtests: verify test data for CR newlines
To detect this without having to run the separate `spacecheck.pl`
script.

It works by passing the new `-w` option. This option is enabled for
all runtests targets except the CI one (to avoid burning CPU cycles
on that for each test run.).

It verifies and warns if the test data contains carriage return bytes.

show the test data filename, minor tweaks