multi admin handle
Treat the multi admin handle more like a "normal" application easy
when it comes to processing and timers. The admin handle is now
always part of the "processing" set (reported "running" counters
do not show this).

Make `multi_runsingle()` have special handling of the admin handle
by calling all tasks it needs to do, like running the connection
shutdown pool. This removes the special "run_cpool" flags during
perform and socket function processing.

This means we can now set other TIMERs on the admin handle as well
and react to it.