GHA: set `--buildinfo` for `test-torture` jobs
For Test Clutch. In an attempt to sync `targetarch` with the rest of
macOS jobs on the feature matrix page:
https://testclutch.curl.se/static/reports/feature-matrix.html

It's `aarch64e` for torture jobs and `aarch64` for the rest.

(stricly speaking aarch64e is the correct value for all jobs, but
autotools and cmake report arm64/aarch64 without the `e`.)

Only the `test-ci` build targets sets `--buildinfo` automatically.
We need to set it manually for other targets used in CI, such as
`test-torture`, to enable the `buildinfo.txt` dump in the runtests step.