speedcheck: do not trigger low speed cancelation on transfers paused with CURL_READFUNC_PAUSE
When a trasfer is paused from a read callback with a CURL_READFUNC_PAUSE code,
it should be excluded from the speedcheck. Currently only transfers paused
from write callbacks are excluded, because the code only checks for
"recv direction" of the transfer. This commit adds a check for "send direction".

Issue similar to https://github.com/curl/curl/issues/6358