add a 'preferred' http version negotiation member
Set 'preferred' on Alt-Svc information against the same destination.
This allows Alt-Svc for h3 to point to h2 and have it tried first. Also,
this allows Alt-Svc to point that http/1.1 is preferred and changes the
oder in the ALPN TLS handhshake.

Add tests in various combination to verify this works.

refs #19874