pytest: quiche flakiness
Let nhttpx only use http/1.1 to backend. This reproduces the
bug in quiche with higher frequency. Allow test_14_05 to now
return a 400 in addition to the 431 we get from a h2 backend
to nghttpx.

Skip test_05_02 in h3 on quiche not newer than version 0.24.4
in which its bug is fixed (https://github.com/cloudflare/quiche/pull/2278)