ws: return specific error for WebSocket failures and cache and reuse connections for WS
Return a CURLE error code that identifies when a WebSocket request fails.  The
return code and http status can be used by the calling application to tell if
the request experienced a refused WebSocket upgrade or not.  The response
headers and body are still available in this case to the application for
processing.  Also the connection is returned to the connection cache for reuse
later if possible.

This change required some extra checks in multi to defer setting the error code
until MSTATE_DONE and to avoid premature close.  I extended the existing
pattern for http2 stream errors in multi.c that prevents close.