tool_doswin: warn user that path is too long to be sanitized
- Treat SANITIZE_ERR_INVALID_PATH in the calling function as
  CURLE_BAD_FUNCTION_ARGUMENT.

- Warn in the sanitizer function via warnf with the reason that
  SANITIZE_ERR_INVALID_PATH is returned.

SANITIZE_ERR_ codes are not mapped directly to CURLE_ error codes which
may have led to some confusion in certain cases so this change is an
attempt to mitigate that.

Prior to this change some calling functions returned CURLE_URL_MALFORMAT
for SANITIZE_ERR_INVALID_PATH even though the URL was not malformatted.

Bug: https://github.com/curl/curl/pull/20113/changes#r2652697794
Reported-by: Viktor Szakats

Closes #xxxxx